small changes to code

+++ labs/06_07_08_test_framework/models/api/admin.py
@@ -28,14 +28,23 @@ class AdminAPI:
     def delete_product_by_name(self, product_name):
         # I did this in my test_products.py, but it seems like doing it here is more reusable,
         # so i simplified the test_products.py using this.
-        response = requests.get(f"{self.base_url}/products", headers=self.headers)
-        products = response.json()
+        #response = requests.get(f"{self.base_url}/products", headers=self.headers)
+        #products = response.json()
         
-        product_id = None
-        for p in products:
-            if p["name"] == product_name:
-                product_id = p["id"]
-                break
+        #product_id = None
+        #for p in products:
+        #    if p["name"] == product_name:
+        #        product_id = p["id"]
+        #        break
         
-        remove_product_response = requests.delete(f"{self.base_url}/product/{product_id}", headers=self.headers)
-        return remove_product_response
+        #remove_product_response = requests.delete(f"{self.base_url}/product/{product_id}", headers=self.headers)
+        #return remove_product_response
+        products = self.get_products()  # returns list of products
+    
+        product_id = next((p["id"] for p in products if p["name"] == product_name), None)
+        if product_id is None:
+            raise ValueError(f"Product {product_name} not found in backend")
+        
+        response = requests.delete(f"{self.base_url}/product/{product_id}", headers=self.headers)
+        response.raise_for_status()
+        return response
